refactor: useControlModalV2 분리
