+++ hook/useControlModal.ts
@@ -1,5 +1,4 @@
-import { atom, useAtom } from 'jotai';
-import { ReactNode, useCallback, useState } from 'react';
+import { useCallback, useState } from 'react';
 
 const useControlModal = () => {
   const [isShow, setIsShow] = useState(false);
@@ -12,33 +11,3 @@ const useControlModal = () => {
 };
 
 export default useControlModal;
-
-const modalAtomInitial = {
-  isOpen: false,
-  content: null,
-  options: { isBottomSheet: false },
-};
-
-const modalAtom = atom<{
-  isOpen: boolean;
-  content: ReactNode;
-  options: { isBottomSheet: boolean };
-}>(modalAtomInitial);
-
-export const useControlModalV2 = () => {
-  const [modalState, setModalState] = useAtom(modalAtom);
-
-  const open = (content?: ReactNode, options?: { isBottomSheet: boolean }) =>
-    setModalState({
-      isOpen: true,
-      content,
-      options: options ? options : modalAtomInitial.options,
-    });
-  const close = () => setModalState(() => ({ ...modalAtomInitial }));
-
-  return {
-    modalState,
-    open,
-    close,
-  };
-};
